requirements.txt
# Core API framework
fastapi>=0.108.0
uvicorn>=0.25.0

# Browser automation (for headless.py)
playwright>=1.39.0

# HTTP requests and data handling
requests>=2.32.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Environment and utilities
python-dotenv>=1.0.0
pyfiglet>=0.8.0

# Type hints support
typing-extensions>=4.8.0